Chore/190 move middleware.ts in to src (#192)
* chore: #190 middleware 파일 이동해서 적용

* fix: #190 lint 에러 수정

* fix: #190 린트 오류 수정

* fix: ci 오류 해결

* chore: trigger ci
